scripts/requirements.txt
python-dotenv
pyotp
requests
psutil
pydantic-settings
